app fully working after frontend restructure routing and directories
