+++ client/App.jsx
@@ -1,18 +1,18 @@
 import React from 'react';
-import DashboardContainer from './components/DashboardContainer';
-import Signup from '../client/components/Signup';
-import Login from '../client/components/Login';
+import DashboardContainer from './components/pages/Dashboard/DashboardContainer';
+import Signup from './components/pages/Signup/Signup';
+import Login from './components/pages/Login/Login';
 import { Route, Switch } from 'react-router-dom';
 import { ProvideAuth } from './routes/useAuth';
 import PrivateRoute from './routes/PrivateRoute';
-import MainNav from './components/MainNav';
+import MainNav from './components/pages/MainNav/MainNav';
 
 const App = () => {
   return (
     <ProvideAuth>
       <MainNav />
       <Switch>
-        <Route exact path="/">
+        <Route exact path="/login">
           <Login />
         </Route>
 
@@ -20,7 +20,7 @@ const App = () => {
           <Signup />
         </Route>
 
-        <PrivateRoute exact path="/dashboard">
+        <PrivateRoute exact path="/">
           <DashboardContainer />
         </PrivateRoute>
       </Switch>
