+++ client/components/pages/Dashboard/Applications/ApplicationsTable.jsx
@@ -1,10 +1,10 @@
 import React, { useState, useEffect } from 'react';
 import { useHistory } from 'react-router-dom';
-import { useAuth } from '../routes/useAuth';
-import ApplicationsTableAddButton from './ApplicationsTableAddButton.jsx';
-import ApplicationsTableHeader from './ApplicationsTableHeader.jsx';
-import ApplicationsTableRows from './ApplicationsTableRows.jsx';
-import SearchBar from './SearchBar';
+import { useAuth } from '../../../../routes/useAuth';
+import ApplicationsTableAddButton from './ApplicationsTableAddButton';
+import ApplicationsTableHeader from './ApplicationsTableHeader';
+import ApplicationsTableRows from './ApplicationsTableRows';
+import SearchBar from '../SearchBar';
 import axios from 'axios';
 
 const ApplicationsTable = () => {
@@ -32,7 +32,7 @@ const ApplicationsTable = () => {
       setUpdateState(false);
     } catch (error) {
       if (error.response.status === 401) {
-        history.push('/');
+        history.push('/login');
       }
       console.log(
         'Error in fetchApplications of DashboardTable component:',
@@ -67,7 +67,7 @@ const ApplicationsTable = () => {
       setUpdateState(true);
     } catch (error) {
       if (error.response.status === 401) {
-        history.push('/');
+        history.push('/login');
       }
       console.log(
         'Error in removeApplications of DashboardTable component:',
