+++ client/components/pages/Dashboard/SearchBar.jsx
