+++ client/components/pages/Dashboard/Steps/StepsTableAddButton.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import StepsModal from './StepsModal.jsx';
+import StepsModal from './StepsModal';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faPlus } from '@fortawesome/free-solid-svg-icons';
