+++ client/components/pages/MainNav/SignoutButton.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useHistory } from 'react-router-dom';
-import { useAuth } from '../routes/useAuth';
+import { useAuth } from '../../../routes/useAuth';
 import axios from 'axios';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faSignOutAlt } from '@fortawesome/free-solid-svg-icons';
@@ -13,11 +13,11 @@ const SignoutButton = () => {
     try {
       const res = await axios.get('/user/logout');
       if (res.status === 200) {
-        auth.signout(() => history.push('/'));
+        auth.signout(() => history.push('/login'));
       }
     } catch (error) {
       if (error.response.status === 401) {
-        history.push('/');
+        history.push('/login');
       }
       console.log(
         'Error in handleSubmit of Logout component:',
