+++ package.json
@@ -31,13 +31,13 @@
     "express": "^4.17.1",
     "jsonwebtoken": "^8.5.1",
     "passport": "^0.4.1",
-    
     "passport-google-oauth20": "^2.0.0",
     "pg": "^8.5.1",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-router-dom": "^5.2.0",
     "react-transition-group": "^4.4.1",
+    "styled-components": "^5.2.1",
     "webpack": "^5.10.3"
   },
   "devDependencies": {
